Separate lints of casts group from types group
